requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
PyJWT
pytest